third_party/WebKit/LayoutTests/fast/frames/frame-dimensions-before-parent-layout.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Tests that querying dimensions in an iframe returns the correct value even before the parent frame has had a layout');

var iframe = document.createElement('iframe');
document.body.appendChild(iframe);

var heightBeforeLayout = iframe.contentWindow.innerHeight;
var forceLayout = iframe.offsetHeight;
var heightAfterLayout = iframe.contentWindow.innerHeight;

shouldBe('heightBeforeLayout', 'heightAfterLayout');

var iframe2 = document.createElement('iframe');
document.body.appendChild(iframe2);

var widthBeforeLayout = iframe2.contentWindow.innerWidth;
var forceLayout = iframe2.offsetHeight;
var widthAfterLayout = iframe2.contentWindow.innerWidth;

shouldBe('widthBeforeLayout', 'widthAfterLayout');
</script>
</body>
</html>
