third_party/blink/web_tests/fast/frames/frame-dimensions-before-parent-layout-expected.txt

Tests that querying dimensions in an iframe returns the correct value even before the parent frame has had a layout

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS heightBeforeLayout is heightAfterLayout
PASS widthBeforeLayout is widthAfterLayout
PASS successfullyParsed is true

TEST COMPLETE
